resolved compile issues
