+++ core/src/gles/context_win32.cpp
@@ -6,6 +6,7 @@
 #include "gles/context.hpp"
 #include "logging.hpp"
 #include "os/surface.hpp"
+#include "psl/stream_utils.hpp"
 #include <Windows.h>
 
 using namespace core;
