+++ psl/inc/psl/stream_utils.hpp
@@ -34,5 +34,35 @@ namespace utility::string
 			ss >> out;
 			return out;
 		}
+		
+		/// \brief cast the input string containing a hex value, to the given type
+		/// \tparam T the type to convert the output to.
+		/// \param[in] str the string containing the hex value.
+		/// \returns an object of type T that has been constructed using the hex value.
+		template <typename T>
+		inline T from_hex(const psl::string8_t& str)
+		{
+			return lexical_cast<T>(str);
+		}
+
+		/// \brief cast the input string containing a hex value, to the given type
+		/// \tparam T the type to convert the output to.
+		/// \param[in] str the string containing the hex value.
+		/// \returns an object of type T that has been constructed using the hex value.
+		template <typename T>
+		inline T from_hex(const psl::string16_t& str)
+		{
+			return lexical_cast<T>(str);
+		}
+
+		/// \brief cast the input string containing a hex value, to the given type
+		/// \tparam T the type to convert the output to.
+		/// \param[in] str the string containing the hex value.
+		/// \returns an object of type T that has been constructed using the hex value.
+		template <typename T>
+		inline T from_hex(const psl::string32_t& str)
+		{
+			return lexical_cast<T>(str);
+		}
     }
 }
